SUppression des Livres

+++ MediaTekDocuments/manager/ApiRest.cs
@@ -96,8 +96,21 @@ namespace MediaTekDocuments.manager
                     httpResponse = httpClient.PutAsync(message, content).Result;
                     break;
                 case "DELETE":
-                    httpResponse = httpClient.DeleteAsync(message).Result;
+                    if (content != null)
+                    {
+                        // Crée une requête DELETE avec body
+                        var request = new HttpRequestMessage(HttpMethod.Delete, message)
+                        {
+                            Content = content
+                        };
+                        httpResponse = httpClient.SendAsync(request).Result;
+                    }
+                    else
+                    {
+                        httpResponse = httpClient.DeleteAsync(message).Result;
+                    }
                     break;
+
                 default:
                     return new JObject();
             }
